ログアウト時にalertを表示するようにする
